src/requirements.txt
flask
logging
PyJWT
redis
pymysql
uwsgi
requests